requirements.txt
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
dotenv==0.9.9
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gcloud==0.18.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jwcrypto==1.5.6
MarkupSafe==3.0.2
oauth2client==4.1.3
protobuf==6.32.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pyparsing==3.2.3
Pyrebase4==4.8.0
python-dotenv==1.1.1
python-jwt==4.1.0
requests==2.32.5
requests-toolbelt==0.10.1
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.43
typing-extensions==4.15.0
urllib3==1.26.20
Werkzeug==3.1.3
